- Add is_installed method in VpnCmdExecutor
