+++ client/python/src/utils/constants.py
@@ -7,6 +7,7 @@ class ErrorCode:
     VPN_SERVICE_IS_NOT_WORKING = 90
     VPN_ACCOUNT_NOT_FOUND = 91
     VPN_ACCOUNT_NOT_MATCH = 92
+    VPN_NOT_YET_INSTALLED = 98
     VPN_ALREADY_INSTALLED = 98
     VPN_START_FAILED = 99
     TIMEOUT = 100
